refactor(ui): rafactor folder structure of binary class
