+++ ui/src/container/models/Details/binary-classification/current/data-drift/list/index.jsx
