+++ ui/src/container/models/Details/binary-classification/current/data-quality/data-point-distribution/index.jsx
