+++ ui/src/container/models/Details/binary-classification/current/data-quality/data-quality-list/categorical/right-table/columns.jsx
