+++ ui/src/container/models/Details/binary-classification/current/data-quality/data-quality-list/numerical/table/index.jsx
