+++ ui/src/container/models/Details/binary-classification/reference/data-quality/data-point-distribution/index.jsx
