+++ ui/src/container/models/Details/binary-classification/reference/data-quality/data-quality-list/categorical/right-table/columns.jsx
