+++ ui/src/container/models/Details/binary-classification/reference/data-quality/data-quality-list/numerical/chart/options.js
