+++ ui/src/container/models/Details/binary-classification/reference/data-quality/index.jsx
@@ -1,24 +1,52 @@
-import { modelsApiSlice } from '@Src/store/state/models/api';
-import { ModelTypeEnum } from '@State/models/constants';
+import SomethingWentWrong from '@Components/ErrorPage/something-went-wrong';
+import JobStatus from '@Components/JobStatus';
+import { JOB_STATUS } from '@Src/constants';
+import { modelsApiSlice } from '@State/models/api';
+import { useGetReferenceDataQualityQueryWithPolling } from '@State/models/polling-hook';
+import { FormbitContextProvider } from '@radicalbit/formbit';
 import { memo } from 'react';
 import { useParams } from 'react-router';
-import BinaryClassificationMetrics from './binary-classification';
+import DataPointDistribution from './data-point-distribution';
+import SearchFeatureList from './search-filter';
+import DataQualityList from './data-quality-list';
 
-const { useGetModelByUUIDQuery } = modelsApiSlice;
+const { useGetReferenceDataQualityQuery } = modelsApiSlice;
 
-function DataQualityMetrics() {
+const initialValues = {
+  __metadata: {
+    selectedFeatures: [],
+    isNumericalSelected: true,
+    isCategoricalSelected: true,
+  },
+};
+
+function BinaryClassificationDataQualityMetrics() {
+  useGetReferenceDataQualityQueryWithPolling();
   const { uuid } = useParams();
-  const { data } = useGetModelByUUIDQuery({ uuid });
+  const { data, isError } = useGetReferenceDataQualityQuery({ uuid });
+  const jobStatus = data?.jobStatus;
 
-  const modelType = data?.modelType;
-
-  switch (modelType) {
-    case ModelTypeEnum.BINARY_CLASSIFICATION:
-      return <BinaryClassificationMetrics />;
-
-    default:
-      return false;
+  if (isError) {
+    return <SomethingWentWrong />;
   }
+
+  if (jobStatus === JOB_STATUS.SUCCEEDED) {
+    return (
+      <FormbitContextProvider initialValues={initialValues}>
+
+        <div className="flex flex-col gap-4 py-4 h-full">
+          <DataPointDistribution />
+
+          <SearchFeatureList />
+
+          <DataQualityList />
+        </div>
+
+      </FormbitContextProvider>
+    );
+  }
+
+  return (<JobStatus jobStatus={jobStatus} />);
 }
 
-export default memo(DataQualityMetrics);
+export default memo(BinaryClassificationDataQualityMetrics);
